// src/core/types.ts
export type ThemeModeTypes = 'light' | 'dark' | 'system'
export type ComponentNameTypes = 'header' | 'sidebar' | 'player'
export type ComponentSkinTypes = 'red' 
| 'green' 
| 'blue' 
| 'orange' 
| 'yellow' 
| 'purple' 
| 'indigo'
| 'pink' 
| 'violet' 
| 'magenta'